Fix imports post-refactor
